requirements.txt
# Core Python Libraries
python-dotenv>=1.0.0

# PDF Processing
PyPDF2>=3.0.0
pymupdf>=1.22.0
pillow>=9.5.0
pytesseract>=0.3.10

# Machine Learning and NLP
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.1.0
google-generativeai>=0.3.0
faiss-cpu>=1.7.4

# Data Handling
pandas>=2.0.0

# Web Interface
gradio>=4.0.0

# Embedding and Vector Store
numpy>=1.24.0

# Optional but recommended
tiktoken>=0.5.0

streamlit
pinecone-client
pinecone-text
pinecone-notebooks
uvicorn
fastapi 
python-multipart
streamlit-feedback
ragas
pymupdf 
pandas 
pillow
# rank_bm25
pytesseract== 0.3.13
# pytesseract 
tabula-py  
# wandb 
# "unstructured[all-docs]" langchain langchain_community \
chromadb 
langchain-experimental
unstructured[all-docs]
onnxruntime
onnx==1.14.0
matplotlib
seaborn
tabulate>=0.9.0
openpyxl>=3.0.0
fpdf
libmagic
python-magic-bin
spacy
-e .